requirements.txt
fastapi[all]
databases
asyncpg
python-multipart
python-jose[cryptography]
passlib[bcrypt]
sqlalchemy
psycopg2
uvicorn
python-dotenv